pyproject.toml
[project]
name = "cursor-secretaria"
version = "0.1.0"
description = "AI Medical Secretary with WhatsApp integration and scheduling capabilities"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "langchain>=0.1.0",
    "langgraph>=0.0.20",
    "sqlalchemy>=2.0.0",
    "google-api-python-client>=2.100.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0"
]
